Move context one level higher
