--- src/client/pages/OriginalDataPoint/components/DataSources/References/AddFromRepository/Props.tsx
@@ -1,4 +1,0 @@
-export type Props = {
-  isOpen: boolean
-  onClose: () => void
-}
