+++ src/client/pages/OriginalDataPoint/components/DataSources/References/Props.ts
@@ -1,0 +1,5 @@
+import { OriginalDataPoint } from 'meta/assessment'
+
+export type Props = {
+  originalDataPoint: OriginalDataPoint
+}
