+++ src/client/pages/OriginalDataPoint/components/DataSources/References/References.tsx
@@ -3,12 +3,12 @@ import { useTranslation } from 'react-i18next'
 
 import { Objects } from 'utils/objects'
 
-import { OriginalDataPoint } from 'meta/assessment'
 import { Topics } from 'meta/messageCenter'
 
 import EditorWYSIWYG from 'client/components/EditorWYSIWYG'
 import MarkdownPreview from 'client/components/MarkdownPreview'
 import ReviewIndicator from 'client/components/ReviewIndicator'
+import { Props } from 'client/pages/OriginalDataPoint/components/DataSources/References/Props'
 
 import { useIsDisabled } from '../hooks/useIsDisabled'
 import { useShowReviewIndicator } from '../hooks/useShowReviewIndicator'
@@ -16,10 +16,6 @@ import { useUpdateDataSources } from '../hooks/useUpdateDataSources'
 import { useEditorOptions } from './hooks/useEditorOptions'
 import { useOnClose } from './hooks/useOnClose'
 import AddFromRepository from './AddFromRepository'
-
-type Props = {
-  originalDataPoint: OriginalDataPoint
-}
 
 const References: React.FC<Props> = (props: Props) => {
   const { originalDataPoint } = props
