--- src/client/pages/OriginalDataPoint/components/DataSources/References/index.ts
@@ -1,1 +1,0 @@
-export { default } from './References'
